General clean of preprocessing (mostly Helpers file creation)
